[Silver III] Title: 주유소, Time: 392 ms, Memory: 36048 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.java
@@ -10,21 +10,21 @@ public class Main {
         int N = Integer.parseInt(st.nextToken());
 
         // 도로 길이 입력 받기
-        int[] roads = new int[N - 1];
+        long[] roads = new long[N - 1];
         st = new StringTokenizer(br.readLine());
         for (int i = 0; i < N - 1; i++) {
             roads[i] = Integer.parseInt(st.nextToken());
         }
 
         // 리터당 가격 입력 받기
-        int[] oils = new int[N - 1];
+        long[] oils = new long[N - 1];
         st = new StringTokenizer(br.readLine());
         for (int i = 0; i < N - 1; i++) {
             oils[i] = Integer.parseInt(st.nextToken());
         }
 
         // greedy
-        int priceSum = oils[0] * roads[0], disSum = roads[0], minPrice = oils[0]; // 초기값 설정
+        long priceSum = oils[0] * roads[0], disSum = roads[0], minPrice = oils[0]; // 초기값 설정
         for (int i = 1; i < N - 1; i++) {
             if (oils[i] < minPrice) { // 최소 가격을 발견했다면
                 minPrice = oils[i]; // 가격 갱신
